tweets/requirements.txt

-i https://pypi.org/simple
aniso8601==9.0.1
attrs==22.1.0 ; python_version >= '3.5'
click==8.1.3 ; python_version >= '3.7'
flask==2.2.2
flask-cors==3.0.10
flask-restplus==0.13.0
flask-restx==1.0.3
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
jsonschema==4.17.0 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
mysql-connector-python==8.0.32;
pyrsistent==0.19.2 ; python_version >= '3.7'
pytz==2022.6
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0 ;
trident-models-package==0.0.1;
werkzeug==2.2.2 ; python_version >= '3.7'